move EntityType to separate file
